+++ apps/web/src/app/[locale]/(admin)/settings/billing/billing-page.tsx
@@ -14,7 +14,7 @@ import {
 } from "@/components/settings/settings";
 import { Trans } from "@/components/trans";
 import { useSubscription } from "@/contexts/plan";
-import { trpc } from "@/utils/trpc/client";
+import { trpc } from "@/trpc/client";
 
 import { BillingPlans, PricingData } from "./billing-plans";
