+++ apps/web/src/app/[locale]/(admin)/settings/profile/profile-picture.tsx
@@ -9,8 +9,8 @@ import { Trans } from "@/components/trans";
 import { useUser } from "@/components/user-provider";
 import { IfCloudHosted } from "@/contexts/environment";
 import { useTranslation } from "@/i18n/client";
+import { trpc } from "@/trpc/client";
 import { usePostHog } from "@/utils/posthog";
-import { trpc } from "@/utils/trpc/client";
 
 const allowedMimeTypes = z.enum(["image/jpeg", "image/png"]);
