+++ apps/web/src/app/providers.tsx
@@ -8,9 +8,9 @@ import { useState } from "react";
 
 import { UserProvider } from "@/components/user-provider";
 import { I18nProvider } from "@/i18n/client";
+import { trpcConfig } from "@/trpc/client/config";
 import { AppRouter } from "@/trpc/routers";
 import { ConnectedDayjsProvider } from "@/utils/dayjs";
-import { trpcConfig } from "@/utils/trpc/config";
 
 export const trpc = createTRPCReact<AppRouter>({
   unstable_overrides: {
