+++ apps/web/src/components/create-poll.tsx
@@ -17,9 +17,9 @@ import { useUnmount } from "react-use";
 import { PollSettingsForm } from "@/components/forms/poll-settings";
 import { Trans } from "@/components/trans";
 import { useUser } from "@/components/user-provider";
+import { trpc } from "@/trpc/client";
 import { setCookie } from "@/utils/cookies";
 import { usePostHog } from "@/utils/posthog";
-import { trpc } from "@/utils/trpc/client";
 
 import { NewEventData, PollDetailsForm, PollOptionsForm } from "./forms";
