+++ apps/web/src/components/poll/manage-poll/delete-poll-dialog.tsx
@@ -10,8 +10,8 @@ import { useRouter } from "next/navigation";
 import * as React from "react";
 
 import { Trans } from "@/components/trans";
+import { trpc } from "@/trpc/client";
 import { usePostHog } from "@/utils/posthog";
-import { trpc } from "@/utils/trpc/client";
 
 export const DeletePollDialog: React.FunctionComponent<{
   open: boolean;
