+++ apps/web/src/components/poll/mutations.ts
@@ -1,6 +1,6 @@
 import { usePoll } from "@/components/poll-context";
+import { trpc } from "@/trpc/client";
 import { usePostHog } from "@/utils/posthog";
-import { trpc } from "@/utils/trpc/client";
 
 import { ParticipantForm } from "./types";
